services/ui-streamlit/requirements.txt
streamlit==1.41.1
requests==2.32.3
python-dotenv==1.0.1
pandas==2.2.2
pyarrow==15.0.2
s3fs==2024.6.1